# Frontend/OpenApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Backend",
    "version": "1.0"
  },
  "paths": {
    "/api/Clientes/{Id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Clientes/MakePurchase": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompraDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Clientes/Login": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDTO"
                }
              }
            }
          }
        }
      }
    },
    "/Clientes/{Id}": {
      "get": {
        "tags": [
          "Compras"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompraDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompraDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompraDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Compras/Filtro": {
      "get": {
        "tags": [
          "Compras"
        ],
        "parameters": [
          {
            "name": "date1",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "date2",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompraDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompraDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompraDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Compras": {
      "get": {
        "tags": [
          "Compras"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompraDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompraDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompraDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Compras/{Id}": {
      "get": {
        "tags": [
          "Compras"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Productos/{Id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Productos": {
      "get": {
        "tags": [
          "Productos"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Productos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClienteDTO": {
        "type": "object",
        "properties": {
          "idCliente": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "contraseña": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompraDTO": {
        "type": "object",
        "properties": {
          "idCompra": {
            "type": "integer",
            "format": "int32"
          },
          "clienteId": {
            "type": "integer",
            "format": "int32"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detalles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DetalleDTO": {
        "type": "object",
        "properties": {
          "idDetalle": {
            "type": "integer",
            "format": "int32"
          },
          "compraId": {
            "type": "integer",
            "format": "int32"
          },
          "productoId": {
            "type": "integer",
            "format": "int32"
          },
          "precioUnitario": {
            "type": "integer",
            "format": "int32"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "subTotal": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductoDTO": {
        "type": "object",
        "properties": {
          "idProducto": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "integer",
            "format": "int32"
          },
          "categoriaId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}